Remove provisional hot-restart cache record-store
Provisional cache record-stores are removed, now we are
creating real record-store during hot-restart. This simplifies logic
in cache hot-restart.

Reasoning behind provisional record-stores was to store reloaded cache
data in a fake record-store and transfer data from this fake record-store
to real record-store when cache is created by user.

But we realized that, this is unnecessary and brings complexity.
`createCache()` looks for registered cache configurations, if there’s no config
and it can register one then it returns success. Since, we are registering
configurations loaded by hot-restart to the service but keeping them
separate, `createCache()` works as there’s no hot-restart involved.