Work around a JIT compiler bug in 1.6u45
The register holding the `address` variable was destructively reused for `pageOffset`
by subtracting `pageBase` from it. This probably happened because, in the code following
the assignment to `pageOffset`, `address` was only used inside an Unsafe call, which 
may have been overlooked by the JIT compiler.

The reordering of the code in this commit forces the `address` variable to be kept 
alongside with `pageOffset`.