Fix a race between read & write of cluster state during hot-restart
Hot-restart status flag is set to `VERIFICATION_AND_LOAD_SUCCEEDED` with a CAS
when verification and load completed successfully. And when CAS is successful,
cluster-state is restored to value read from disk.

When another process reads hot-restart status as `VERIFICATION_AND_LOAD_SUCCEEDED`,
it also reads current cluster-state but since write of hot-restart status and
cluster-state is not atomic, it can read an inconsistent state.

To avoid that, added another flag `finalClusterStateSet` and while reading
cluster-state, reader process will spin until it observes that flag as set if
hot-restart status is `VERIFICATION_AND_LOAD_SUCCEEDED`.