Hot Restart should fail when an unexpected member joins
Each member knows the latest members list of cluster before shutdown
and expects all or a portion of them to join back.
If an unexpected/unknown member is detected during validation phase,
hot restart fails immediately, because it's obvious that these members
are not the part of the same cluster.