GlobalPoolingMemoryManager does not rely on provided offset during block initialization
Reasoning:
There is a data race between offset acquisition and block initialization.
It's very well possible the offset was read while another thread was
mutating the block.

The markInvalid() method in the GlobalPMM already ignores the provided offset.

As a general rule we should not trust informations stored in a block unless
we successfully marked the block as unavailable for others.
In other words: "You can't trust it unless it's yours."