Fix hot backup interrupt and failure handling
Fix failure handling for hot backup in case there is an existing in_progress file in the target directory. Previously the backup continued after detecting this issue. Hot backup operation was failing on the sending side because it was sending the operation to itself and on the receiving side since the operation didn't have the callId set and expected a response.
Increased code coverage for hot backup feature.

Fixes : https://github.com/hazelcast/hazelcast-enterprise/issues/1221