Add WAN operation heartbeats
`WanOperation` and `WanSyncOperation` offload their work to a different thread but do not send operation heartbeats. On the sender side, this causes operation-heartbeat exception. The publisher and wan service now keep track of these ongoing operations and send heartbeats to the sender. The operation can still timeout if configured so but not because of missing heartbeats.