Respect partition table assignments on `EWRRemoveBackupOperation`
When the wan event is successfully sent to the target cluster, the backups are removed from the replicas. Problems arise if the structure is defined to have more replicas then partition groups. Currently this was checked by sending to less replicas than there are cluster members but this is not the case when the cluster has a partition group type other than PER_MEMBER. In that case there can be more cluster members than partition groups. Thus, the operation is sent to a replica which doesn't exist. The fix checks the max backup count rather than the cluster size before sending the operation.

Fixes : https://hazelcast.zendesk.com/agent/tickets/2402