Clients must observe membership changes before starting quorum tests
The test initialisation code splits the cluster into 3+2 members and waits until the members observe these changes (both in member list and quorum results). But it does not wait until the clients observe the new cluster formation. The clients can then still believe that the cluster is split into 4+1 or even think the cluster is not split yet (as the membership change is asynchronous). Then it can send an invocation to a member which will later be removed from its member list, causing:
- exception on invocation since the membership change throws an exception on existing invocations to removed members
- false quorum results - it expects one result but sends the operation to the wrong sub-cluster
The fix waits until the clients observe the membership change as well before starting the test.

Fixes:
https://github.com/hazelcast/hazelcast-enterprise/issues/1352
https://github.com/hazelcast/hazelcast-enterprise/issues/1351
https://github.com/hazelcast/hazelcast-enterprise/issues/1339
https://github.com/hazelcast/hazelcast-enterprise/issues/1332
https://github.com/hazelcast/hazelcast-enterprise/issues/1322
https://github.com/hazelcast/hazelcast-enterprise/issues/1321
https://github.com/hazelcast/hazelcast-enterprise/issues/987