Removed synchronized block in SSLSocketChannelWrapper
Only during handshaking synchronization is needed, but during normal
operation the only concurrency constraint is that no concurrent wraps
are allowed. Or concurrent unwraps are allowed.

But this isn't possible because a single thread handles the reading part
and a single thread handles the writing part.

So once the handshaking is complete, synchronized is not needed in our design.

For more information see:
https://docs.oracle.com/javase/7/docs/api/javax/net/ssl/SSLEngine.html
And scroll down to 'concurrency notes'