Simplified classes with HiDensity and NativeMemory
So instead of having a class with both HiDensity and NativeMemory,
we just keep the specialization 'NativeMemory'.

It is like naming with animals. A Dog is a type of Mamel,
we don't say DogMamel.

The same goes for e.g NativeMemoryNearCacheRecord.
This is a type of HiDensityNearCacheRecord, so we don't need to
name it HiDensitityNativeMemoryNearCacheRecord.

The problems with these long names is that they are so long, it
because very difficult to try to understand the code.