Add WAN publisher support for discovery SPI
Change WAN connection manager, publisher configuration and
implementation  to support discovery SPI. The current mechanism of
defining endpoints is turned into an implementation of the discovery
SPI. The connection manager now reruns discovery on a periodic basis
and adds the discovered nodes to the available list. The size of the
list can be limited.

Notes:
- the striped executor in the WanBatchReplication is sized according
to the number of target endpoints when sending the first event and
does not resize afterwards. There is a configuration option to set
the executor thread count manually.
- the AbstractWanReplication implementation of isConnected has
changed - previously it checked that there were no failed connections,
now it checks if there are any discovered endpoints
- the periodic discovery task does not attempt to connect to the
target endpoint before adding it to the list