Fix ForceStartTest failure
A node should be able to perform a graceful restart after a force start. This test is verifying that behaviour. However, we don't allow graceful restarts on ACTIVE cluster state. Even if the partition assignments are not done, we consider it as an internal detail and apply the general rule.

Fixes https://github.com/hazelcast/hazelcast-enterprise/issues/1424