WAN operations return EnterpriseWanReplicationService as service
WAN operations should ideally return MapService or CacheService,
depending on which service the event belongs to. But that also means
that the operations should be service namespace aware and that
WanReplicationService should be FragmentedMigrationAwareService and
return replication data for map or cache when the relevant map or
cache partition namespace is being replicated/migrated.

This fix temporarily fixes the tests by returning
EnterpriseWanReplicationService as the service which then causes all
WAN operations and backup operations to belong in the internal
(non-fragmented) namespace.

Fixes:
https://github.com/hazelcast/hazelcast-enterprise/issues/1495