Improve WAN connection handling
There are several minor improvements for the connection handling:
- previously WAN failed to start when on startup it didn't have any
endpoints to connect to. This was ok because the endpoints were
defined in the configuration and there was no way of adding new
endpoints at runtime. Now with discovery, WAN can start up without
any endpoints and they can be added via discovery at a later point.
We leave a warning as without endpoints the WAN queues will fill up
- remove "live" endpoints if they are not returned by the discovery
SPI. Previously we left endpoints which we know are alive but were
not returned by the discovery SPI. This could happen, for instance,
if the node no longer passed the discovery SPI filter (e.g. it's tag
was changed). Now the target endpoint list and discovery SPI list are
 in sync
- remove endpoints to which we cannot connect (they are unreachable).
Previously they were kept in the list meaning that they were retried
indefinitely. Since WAN already removes endpoints for which the
connections went dead and for which we received an exception on
replication, additionally removing an unreachable endpoint is fine.
The endpoint can be re-added later if it is rediscovered.