Improved TLS exception in case of TLS handshake failure.
In the old code the socket.read result isn't checked for being -1;
indicating the socket is closed.

So eventually the handshake gives up after 250 attemps and will give
an obscure error.

This PR changes that by explicitly checking for -1 and explaining what
has happend and what is likely to be the cause of the problem.