Fixes OpenSSL santity check problem
Fix #1601

The problem was that the sanity check created the SSLEngine
in client mode and in server mode. But for a java-client, starting
the engine in server mode makes no sense since the server mode
for example requires a keystore containing the certificate of
the server; for a java client there is no such need.