Make WanReplicationService a FragmentedMigrationAwareService
WAN operations should ideally return MapService or CacheService as
the service name of the operation, depending on which service the event
belongs to.
But that also means that the operations should be service namespace
aware and that WanReplicationService should be
FragmentedMigrationAwareService. The WanReplicationService then
returns replication data for map or cache when the relevant map or
cache partition namespace is being replicated/migrated.

The fix also :
- reverts the temporary fix which then caused all WAN operations and
backup operations to belong in the internal (non-fragmented)
namespace
- refactores the EnterpriseWanReplicationService by
delegating some code to the WanReplicationPublisherDelegate instead
of inspecting it's fields.

Fixes:
https://github.com/hazelcast/hazelcast-enterprise/issues/1495
https://github.com/hazelcast/hazelcast-enterprise/issues/1596