Harden RedBlack from NOOME inconsistencies (#1584)
When adding a new Node in the Tree, we allocate the Node size, and attach
the new address as a leaf to the parent Node. Any allocation taking place
after we make this link, if not successful, can leave the structure with
a reference to an incomplete Node.
With this patch, we only promote the Node as a leaf, once every allocation
was successful.

A few more improvements introduced for better debugging.