Fix leak of offloaded operation set in WAN service
There was a race that could lead to an operation first being
processed in a different thread and then being added to the set of live
operations (to notify the invocation monitor the operation is being
processed in a different thread). This caused a memory leak.

The fix first adds the operation and then offloads to a different
thread. Extra care is taken to log warnings in case the operation
wasn't removed from the set or that the operation is removed if the
offloading failed.