Race on SSLChannel.sslEngineResult
The problem is that wrap and unwrap both read/write the sslEngineResult.

This is done for the handshaking process; but it is also done for dealing with the
result of a wrap/unwrap. So it could be that a unwrap is done and the sslEngineResult
is set (and expected not to change), the wrap comes through and modifies the sslEngineResult,
and then the unwrap continues and now is going to process the sslEngineResult from the wrap!

The problem is solved by introducing two local variables; wrapResult and unwrapResult. These
can't be modified by another thread since they are local variables and the wrap/unwrap
methods work with these local variables. The sslEngineResult has been renamed to handshakeResult
and is being published by wrap/unwrap.

Because the handshake is done in a synchronized block there is no concurrent access to this
handshakeResult, so no races here.