Minor SSLEngine cleanups
These cleanups are pulled out to seperate it from a bug fix pr which is pending.

Renamed some arguments

Moved the wrap/unwrap methods close to each other.

Removed unused return value for unwrap

Removed useless arguments being passed which are alwas pointing to the same fields.

This change is backported, so that the F5/Netty PR:
https://github.com/hazelcast/hazelcast-enterprise/pull/1711
Can be merged in easily.