Initialize WAN publisher when checking if WAN is enabled
The WAN publisher was initialized when the proxy was being created.
There was a race condition here when a proxy was created on the
backup replica and the operation was then sent to the partition owner.
It could happen that the operation was received by the partition
owner before the proxy was initialized. In this case, the WAN
publisher was not initialized either and WAN was reported to be
disabled. This could cause a WAN event to be lost.

The fix initializes the WAN publisher when checking if WAN is enabled.
This is possible because the operation that invokes the check has
already created the cache config.

Also, renamed parameters for some cache related methods and added
javadoc to document which parameters were the prefixed cache name.

Fixes : https://github.com/hazelcast/hazelcast-enterprise/issues/1572