Fix atomicity issue in hot restart member list persistence
Member list changes are persisted asynchronously. During shutdown, although shutdown procedure persists the most recent member list, a stale member list could be persisted due to lack of atomicity in check-then-act behaviour of onMembershipChange() method.

With this fix, member list changes are persisted synchronously. When shutdown is triggered, there will not be any pending stale member list update to be persisted.

There is another fix for a minor issue. If force start is decided just when a member times out on member list validation, although the member is able to detect that it needs a force start, it immediately fails hot restart process by first checking the timeout. This issue can occur easily if the local environment is slow. To solve the problem, when timeout occurs, a node can first check the global hot restart result and act accordingly if hot restart result is present.

Fixes #1577
Fixes #640