Validate WAN endpoint configuration
When using the endpoints property in the hazelcast configuration and
the property has been misconfigured, the node will fail to start with a
InvalidConfigurationException. When using some discovery SPI
implementation, the node will start with a warning as we assume the
discovery SPI is currently not working but will possibly work later.

Improved the error message when the endpoints are misconfigured.

Fixes : https://github.com/hazelcast/hazelcast-enterprise/issues/1615