Log slow disk IO while persisting Hot Restart cluster metadata
Currently member list and partition table are persisted while holding
their service level locks. This is to provide ordering of write events.
We can offload persistence to a specific thread but we are not following this
path for the sake of simplicity. Because, we assume Hot Restart will be used
with a fast(?) SSD and metadata will be relatively small in size.

Logging slow disk persistence will give us more hints.