Skip disposing deferred blocks in case of partition agnostic operation
All operations extending HDMapOperation dispose deferred blocks on
error and as such are expected to be partition operations. This is
not always the case. For instance, the HDQueryOperation is run on a
generic thread. The fix skips disposing deferred blocks in case when
the operation is run on a generic thread.

Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/1735

(cherry picked from commit 97c8ff0)