Disable event journal when cluster version is less than 3.9
Even though the event journal is configured as enabled, we decide to
ignore this setting as there might be some pre-3.9 members in the
cluster which don't know how to handle event journal data. One option
is to keep the journal data on 3.9 members and lose the data once
that partition is migrated to 3.8 members but this is more confusing.
Since we don't allow subscribing to or reading from the event journal
when the cluster version is not 3.9 or higher, not saving to it is
completely fine.