Add configuration property for WAN to use private endpoint address
When using the discovery SPI, WAN by default uses the public address of
the discovered endpoints. In some cases, though, the user may wish to
use the private address (such as when the instances do not have a public
address at all).
Also, did some refactoring to increase the separation of concerns and
added some javadoc for clarity. Added a null check when adding
discovered endpoint addresses to avoid a NPE later in the execution.

Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/1776