WAN snapshotEnabled fixes
- when snapshotEnabled was enabled, the the code coalesced events for
the same key but on different structures (different maps and/or caches).
We introduce a composite key which contains the service name, object
name and entry key for coalescing to avoid the issue
- when snapshotEnabled was enabled, the WAN queue size counter was
decremented by the wrong number of processed events and did not count in
 the coalesced events
- added a check that an empty endpoint defined in the configuration
isn't added to the list of valid WAN endpoints
- refactored the BatchWanReplicationEvent to avoid creating
collections which aren't used and made the BatchWanReplicationEvent more
explicit. Previously the code used some subtle assumptions that involved
calling a getter exactly once.

Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/1798