The problem was that the wrap method ended up in a loop. When sslengine.wrap is called, is signals there is not enough space for writing.
Then the sslsession is asked how big the buffer should be. And if the session
says the buffer isn't big enough, the buffer is expanded.

The problem is when the sslsession packetbuffer size equals the actual buffer capacity
but the buffer is dirty. Then there is not enough space to encode, but there is also no
reason to increase the capacity of the buffer since the buffer capacity equals
session.packetBufferSize.

The consequence is that the io thread gets into a loop and therefor will not process
any other requests and eventually the connections become useless and get closed.

The fix: as soon as the sslengine wrap indicates there is no space and the buffer capacity
equals the session.packetBufferSize, the wrap is ended and system flushes the the content
of the buffer to the socket. Because not all bytes are send, the sslchannel will register
for writing, will immediately get a signal there is space and then continues the
encoding process but now with an empty buffer and this time the wrap will succeed.

The broken behavior was introduced in this PR: #1718