Use serial runner for the hot restart cluster start tests (#1954)
The hot restart cluster start tests rely on timeouts to test specific scenarios. This requirement makes them fragile to liveliness issues in the build environment. To make them less fragile, these tests are updated to use the serial test runner. In addition, PartialStartTest is completely moved to the SlowTest category.