Added generics to SplitBrainMergePolicy
The generics on the SplitBrainMergePolicy interface, make it clear
which merge types a merge policy requires.

For example the HigherHitsMergePolicy<V, T extends MergingHits<V>>
requires the MergingHits type to be provided by the data
structure. This prevents the downcasting to that interface, without
knowing if the type is supported.

The merge types can even be composed like <V, T extends
MergingHits<V> & MergingCosts<V>>, so you can write very specific
merge policies in a very simple way.