Change order of WAN initialisation
Changed order of WAN initialisation to first parse the configuration
properties before scheduling tasks, calling external API (e.g. AWS API
with discovery SPI).
Previously, because of the class hierarchy, some the configuration was
parsed after the discovery SPI was invoked and after a task has been
scheduled. If the configuration parsing threw an exception, this would
waste an invocation to the discovery SPI and would not clean up
resources. The fix moves all configuration setup to the beginning of
each init() method.

Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/1935