Converted HDQueryOperation to new batch functionality. (#2066)
The old batch functionality was problematic since it runs
all partitions for a partition thread without the ability
to interleave other operations. This leads to very bad
response times if e.g. a map.get is done while a hd query
is being executed.

With the new batch functionality, this problem is gone.