Fix racy MapWanReplicationQuickTest.testClearQueues
The issue is caused by a race in the test. It may happen that events are drained from the WAN and staging queues and are already waiting for transmission when clearQueues is invoked. In this case, the queues are empty, so that nothing can be cleared, while the counter still says that there are 1000 events in the memory. The issue is reproducible with adding a few secs sleep right before the clearQueues call.

The race has been solved by pausing the replication before events are populated to the WAN queues, so that the events are still there and can be cleared by the `clearQueues` call.

Fixes #2117