Fixed a wrong assertion method in a WAN test
`checkDataInFrom()` actually does not enforce anything. It just returns
false when data from one cluster is not present in the other one.

`assertDataInFrom()` does the actual assertion.

Ideally the `checkDataInFrom()` should be renamed to something
`isDataInFrom()` to clearly indicate it just returns the state.
However this WAN test suite is used in the Solace project and it will
require a little of coordination.