Fix `VTopo_1passiveReplica_2producers_withPutIfAbsentMapMergePolicy` test
Reasoning:
The test was simply badly designed. It was passing (most of the time) just
by coincidence & WAN event coalescing.

Here is the original scenario:
It would insert a `dummy` value on a source side and this `dummy` value
was immediately overwritten by `another` value. Then it would assert the
`another` value made it into the target side.

The test uses the `PutIfAbsent` merging policy. So when the `dummy` value
was coalesced with the `another` value on the source side (=most of the time)
then the test would pass. As the target cluster received just the `another`
value.

However when the `dummy` value made it to the target side before the
source side had a chance to coalesce it with the `another` value then
the test failed. It's because the target side received the `dummy` value
and then it was never rewritten by the `another` value. The `another`
value was ignored due the `PutIfAbsent` merging policy.

Fixed #2134