Eliminate race condition in WanDiscoveryTest test assertion
A WAN endpoint can only be removed from the endpoint list as a result of
exceptions when replicating WAN events. In a quiescent state, WAN
endpoints which are not reachable will not be removed. The second part
of the test relies on this by generating WAN events after shutting down
a target endpoint. As a result, the endpoint will be removed and the
following assertion may pass. Unfortunately, concurrently the discovery
implementation is run each second. As it does not know that the target
endpoint has been shut down and as it always returns the same set of
nodes, it will effectively add the shutdown endpoint back into the
target endpoint list. Once the endpoint is added, it will not be
removed, since there are no more generated WAN events. This is why the
test assertion might fail.
The fix actually waits until there are again 2 endpoints in the target
endpoint list. As described, there will temporarily be only 1 endpoint
in the list, but this is not always observeable and we are better off
waiting until the list is back to 2 endpoints.

Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/2148