Introduce a common ancestor for Map/Cache HotRestart tests
Reasoning:
There are multiple tests dealing with the same concerns:
creating a unique temporary directory, deleting it, restarting cluster, etc.

Having this logic copy&pasted is error-prone as shown by #2191

There are still some advanced test which do not use the newly introduced
support, but they are not not testing HotRestart any HR-based
data-structure, but they test HotRestart itself - these tests
are too specialized and I am not touching them in this changeset.

Future Work: Move the HR-specific logic into a junit rule rather than
having it as a superclass. Rules compose better and we already use
this approach for bouncing tests.