Fix EP index queries consistency
EPs equipped with predicates are trying to utilize global indexes to
narrow down the key set to operate on. This change fixes the following
problems:

1. There were no migration stamp validations performed to prevent EPs
from operating on stale index data (original fix and idea by Matko).

2. We were not producing WrongTargetException for out-migrated
partitions, as a result the caller (InvokeOnPartitions) was not retrying
EPs on such partitions.

Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/2277