Scan fixed number of entries to find expired ones.
For HD Near Cache, this scanning is done under lock
and it was a full-scan on each segment. This had a negative
impact on performance of other methods that were waiting to
acquire lock. With this fix, we will start to scan only a
fixed number of entries in each loop.