Call MerkleTree update methods with Data value paramerer
`MerkleTreeUpdaterRecordStoreMutationObserver#onUpdateRecord` receives `Record` and `HeapData` as old and new values respectively, which use different `hashCode` implementations in `OBJECT` in-memory format: `Object.hashCode()` and `HeapData.hashCode()` (Murmur hash). As a result, replacing a record to the same value results in changed Merkle tree hashes in the branch of the updated record, which is later reported as an incostistency by the Merkle tree check.
This commit fixes this behavior by converting all objects that the `MerkleTree` is called with to `Data` by using `SerializationService`.