Serialize node values in MerkleTreeNodeCompareOperation + minor changes
Missing serialization of the remote node values in the
`MerkleTreeNodeCompareOperation` may mark some identical partitions as
partitions with differences, which eventually leads to syncing more
partitions than needed. This happens if the operation is serialized in
the target cluster.

The commit includes some other minor changes:
- Stronger assertion in `WanMerkleAntiEntropyTest` with concrete
expectation about the number of partitions with differences instead of a
wide range.
- `EnterprisePartitionContainer#getMerkleTree` should be called only
during map context creation, others should use `getMerkleTreeOrNull`
since the first one is expensive. Renamed the method to
`getOrCreateMerkleTree`.

Fixes #2348