Use constant key for remove in WanMerkleAntiEntropyTest
The test expects that after the removal the Merkle consistency check identifies a single Merkle tree leaf as different, the one holded only the removed entry, therefore it expects no record needs to be synchronized. This expectation fails if the Merkle tree leaf is shared by more then one entry. With random remove keys we may remove an entry from a shared leaf. This change fixes the flaky behavior of the test by using a constant key.

Fixes #2500