Fix WAN counters behaviour on migrations (#2499)
Fix WAN counters behaviour on migrations

During migrations, WAN events were counted according to the partition
table as it would be after migration commit (future primary events were
counted as primary, future backup events as backups).
In addition to this, in WanQueueMigrationSupport#onMigrationCommit,
event counts would then be transferred from a backup to a primary
counter and vice versa. At this stage, events enqueued during a
replication operation should not be included in this transfer of
counted events but they were and counters could go negative.
The fix makes the replication operation count the events according to
the previous partition table (before the commit), which causes the
commit to correctly transfer the counts of all events.
In case the member was not a replica, the replication operation will add
them as backup events and the migration commit will treat them as backup
events.

Thanks to @blazember for the reproducer and the ideas.