Use constant key for update in WanMerkleAntiEntropyTest
The test expects that after the update the Merkle consistency check identifies a single Merkle tree leaf as different, the one holded only the updated entry, therefore it expects only one record needs to be synchronized. This expectation fails if the Merkle tree leaf is shared by more then one entry. With random keys we may update an entry in a shared leaf. This change fixes the flaky behavior of the test by using a constant key.

Besides fixing this issue, this commit changes the random key in an ignored case too, in order to prevent potential failures in the future when enabling the ignored case.

Fixes #2514

Identical fix as it was in #2505