Fix WAN counter issues with snapshot enabled
If snapshots are enabled, the Batching sender may coalesces multiple events for the same IMap entry. After sending the event to the target it sends an `EWRRemoveBackupOperation` to remove the backups, but it does it only once for one entry despite the number of the events coalesced. This may make the nodes holding the backups to keep the coalesced events in their queues and potentially replicating them if there is a migration. Such a case - besides the broken counters - may violate monotonic writes on the target. Consider the following WAN queue for a certain partition: `(K;Z)`,`(K;Y)`,`(K;X)` with head on the right. These events are getting coalesced as `(K;Z)`, and the now the target has `(K;Z)`, which is correct. Then only one `EWRRemoveBackupOperation` is sent for the three coalesced events and the backup queue remains with `(K;Z)`,`(K;Y)` in it. At this point, if a migration occurs, the queues are replicated and with unlucky batch boundary - the two values belong to the same key are not getting coalesced - the new owner may send `(K;Y)` to the target.

This fix maintains the coalesced events in a temporary structure in the `BatchReplicationEvent` and sends an `EWRRemoveBackupOperation` for each of them.

Besides the prod code change the tests are updated as well.