Fix NPE in WanPublisherFullSyncSupport when syncing all maps
Syncing all maps doesn't work as expected:
- While syncing it tries to put into a map with key `event.getMapName()`, which is `null` in this case throwing a NPE
- Because of the NPE, the synchronization doesn't finish, not sending the result to the caller, operation is not removed from the `liveOperations`
- One sync statistics object is maintained for all synchronized maps

Despite these above, the synchronization is properly done - the clusters got synced - hence the tests didn't catch the issue.

This behavior is fixed by getting the maps to synchronize from the MapContainer in the sync all maps case - instead of getting it from the event - and by maintaining WAN sync statistics separated for each map.
Sending events to MC is also changed to per map basis.

`MapWanSyncAPITest` is changed to verify if the statistics are updated for each map, meaning the synchronization finished successfully.

Fixes #2523