Implement automatically removal of stale Hot Restart data
Currently, when a member is restarted with Hot Restart feature is enabled,
then reconnecting member to the cluster fails unless cluster is in one of
FROZEN or PASSIVE states. User has to remove manually the hot-restart member data
to be able to join to the cluster as a fresh member. But in most cases,
this is a burden on Ops people and manual operation is not cloud friendly.

As a solution, we'll give an option to automatically remove hot-restart data
when cluster is in ACTIVE state. Since we know that partitions are already rebalanced
and data is migrated between existing members. Data persisted in restarting member's
storage is stale and safe to remove.