Track indexed partitions to avoid unnecessary reindexing (#2543)
This change fixes the following issues:

1. When a user invokes IMap.addIndex, AddIndexOperation is sent to all
members and the index in question is fully rebuilt even if it's already
populated. Typically, such addIndex invocations are performed from user
code for each member joining a cluster during the member initialization.
If the cluster is storing a lot of data, rebuilding all of the indexes
may be a very costly operation.

2. When a new member joins a cluster, a local map proxy object is
created on that new member. During the proxy initialization,
AddIndexOperation is sent to all cluster members for each index
configured by the new member XML configuration to ensure the indexes
are present on all members (see MapProxySupport.initializeIndexes).
Basically, we are rebuilding all of the indexes on all of the members.

This change tracks a set of partitions indexed by each index. If a
partition is already indexed by a certain index, the reindexing is
skipped.

Fixes: https://github.com/hazelcast/hazelcast/issues/13964