Improve execution time of MapWanBatchReplicationTest
MapWanBatchReplicationTest.mapWanEventFilter_prevents_replication_of_loaded_entries_by_default
needs the most time from all EE tests running under the default-test
goal to complete: 2m+ constantly.

Since the test class is parameterized, it occupies a JVM for minutes
while not doing anything just waiting for a condition to be held for
minutes. This condition verifies that we don't replicate IMap entries
loaded from the MapStore. This behavior can be verified in the source
cluster by configuring the replication with PAUSED state and querying
the WAN outbound counter directly from the WanBatchReplication instance,
which counter is updated synchronously on the partition operation thread
whenever an operation attempts to publish an event over WAN. This change
gives an improvement in execution time, by reducing the time from 2m+ to
~2s in local.

Besides the change above, this test case and the similar
mapWanEventFilter_allows_replication_of_loaded_entries_when_customized
is configured to run with single-node clusters further reducing the
overhead.