Fix for hdlocalMapStats
Owned entry count test is seperated.
Because it creates the cluster with different config, it can not
work with base class anymore.

Query operations(keyset, values, entryset) seem to be updating]
the stats per partition because of offloading, it seems more
appropriate the fix the test in this case. Tests are ignored
for now.

related to #2704