Added non-blocking concurrent WAN replication mechanism (#2667)
Added non-blocking concurrent WAN replication mechanism

Refactored WanBatchReplication to allow concurrently sending batches of
events for the same partition. This allows greater throughput for use
cases where the events for a single partition are commutative, e.g. only
adding new and immutable entries. Throughput for the default case where
there is only one batch per partition being sent concurrently has also
improved as a side-effect.

Some big changes introduced by this PR:
- a WAN executor was introduced to process WAN replication tasks. By
default the cached executor is used but the user can configure a
concrete executor
- there is no more staging queue. The thread which dequeues events from
WAN replication queues is the same thread that actually transmits the
batch to the target endpoint
- there is an auxiliary queue for WAN sync events. This queue can only
contain WAN sync events for full and merkle-tree sync. It is filled from
 a thread processing the anti-entropy request on the WAN executor
- the WAN sync events are sent randomly to target endpoints and no
ordering is attempted for WAN sync events. This is because the WAN sync
events are strictly independent add events and as such are commutative.
Reordering with other WAN events from WAN replication queues is also
acceptable at this point as there was no ordering even for the previous
implementation
- events are no longer dequeued from WAN queues by polling one-by-one
but rather by draining up to the batch size. This may decrease fairness
of collecting events from different partitions but it has several
benefits: it will potentially increase throughput since WAN backup
events are mostly contained in less partitions, the target cluster
can more efficiently process events if they are contained in less
partitions and dequeueing by draining will introduce less contention on
the WAN partition queues
- no new tests were introduced thus far. Existing tests were
parametrized as all of the test scenarios should pass with both the
existing and the new implementation

This is a base PR for further improvements to make reviews easier.
Follow-up PRs will introduce more improvements (in order of preference):
- removal of WAN event backups can be improved by batching events and
sending the number of replicated events instead of the actual keys and
values in AbstractWanPublisher.removeReplicationEvent. This method may
also be turned into non-blocking
- on the target cluster, the events can be processed concurrently for
different partitions (see
EnterpriseMapReplicationSupportingService.onReplicationEvent and
CacheReplicationSupportingService.onReplicationEvent)
- the batch collection can remember the last partition ID from which
events were collected and continue collection from there. This is to
remove bias for collecting events from partitions with lower partition
IDs
- further diagnostics can be introduced
- each invocation of PartitionWanEventContainer.drainRandomWanQueue will
create a new iterator on the map values. This generates garbage and
introduces contention and as such was always present in the JFR
reports. The structure of PartitionWanEventQueueMap may be changed to
improve the random-access pattern as the most common use-case. Other
scenarios such as migration may become less efficient but this is
acceptable as they are less frequent
- the WanBatchReplication.collectEventBatch method can be simplified
- different event counts in BatchWanReplicationEvent can maybe be
simplified
- other simplifications and javadoc